pyproject.toml
[project]
name = "cxo-agent"
version = "0.1.0"
description = "Semi-automated pipeline for the meerkat-PhD work-and-productivity Instagram account"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.39",
    "httpx>=0.27",
    "feedparser>=6.0",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "rich>=13.7",
    "pyyaml>=6.0",
]

[project.scripts]
cxo = "cxo_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cxo_agent"]
